# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Urja Meter API",
    "description": "\nA clean, read-only REST API over the **Urja Meter Ops** portal (smart-meter data for a\nJaipur distribution utility). The portal has no API of its own. This service logs in as a\nnormal user, reads the portal's internal endpoints (see `PROTOCOL.md` in the repository)\nand serves the data as typed JSON.\n\n* **Meters**: nameplate, status, location, network position; filter by any attribute,\n  network node or distance.\n* **Readings / consumption**: register readings normalised to numbers and ISO timestamps\n  (IST), with per-interval and per-day consumption derived from register deltas.\n* **Network**: the zone \u2192 circle \u2192 division \u2192 subdivision \u2192 substation \u2192 feeder \u2192\n  transformer hierarchy, reconstructed and repaired from per-meter data.\n* **Data quality**: every value we repaired or could not reconcile, with the portal's\n  original value.\n\nReference data is served from a local index synced from the portal's bulk export;\nreadings are fetched on demand and cached. `GET /v1/status` shows how fresh everything is.\nErrors use `application/problem+json` (RFC 9457) with a stable `code` (listed under the\n`Problem` schema). Unknown query parameters are rejected rather than ignored.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/transformers/{code}": {
      "get": {
        "tags": [
          "transformers"
        ],
        "summary": "Get one transformer with its meters",
        "operationId": "get_transformer",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Transformer code (any case).",
              "examples": [
                "DT-001"
              ],
              "title": "Code"
            },
            "description": "Transformer code (any case)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformerDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/network": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Shape of the reconstructed network",
        "operationId": "network_overview",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkOverview"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/network/tree": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "Drill-down tree of the network",
        "description": "Zone \u2192 \u2026 \u2192 transformer, built from the transformers' paths. A node means \"this\ncode under this path\", so a code reused under several parents appears once under\neach, and the counts at every level add up to the fleet total.",
        "operationId": "network_tree",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 7,
              "minimum": 1,
              "description": "How many levels to return, from zone down.",
              "default": 7,
              "title": "Depth"
            },
            "description": "How many levels to return, from zone down."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkTreeNode"
                  },
                  "title": "Response Network Tree"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/network/{level}": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "All nodes at one level",
        "operationId": "network_level",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NetworkLevel",
              "description": "A level of the network, top to bottom."
            },
            "description": "A level of the network, top to bottom."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NetworkNodeSummary"
                  },
                  "title": "Response Network Level"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/network/{level}/{code}": {
      "get": {
        "tags": [
          "network"
        ],
        "summary": "One network node",
        "operationId": "network_node",
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NetworkLevel",
              "description": "A level of the network, top to bottom."
            },
            "description": "A level of the network, top to bottom."
          },
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The node's code at that level (any case).",
              "examples": [
                "DT-007"
              ],
              "title": "Code"
            },
            "description": "The node's code at that level (any case)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkNode"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/data-quality": {
      "get": {
        "tags": [
          "network",
          "data quality"
        ],
        "summary": "Everything we had to correct or could not reconcile",
        "operationId": "data_quality",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataQualityReport"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataQualityReport": {
        "properties": {
          "meters_with_issues": {
            "type": "integer",
            "title": "Meters With Issues"
          },
          "issues_by_code": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "blank_code",
                "blank_name",
                "stale_name",
                "path_conflict",
                "unrecognised_value"
              ]
            },
            "type": "object",
            "title": "Issues By Code"
          },
          "meters": {
            "items": {
              "$ref": "#/components/schemas/MeterIssues"
            },
            "type": "array",
            "title": "Meters"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkOverview"
          },
          "readings": {
            "$ref": "#/components/schemas/ReadingsQuality"
          }
        },
        "type": "object",
        "required": [
          "meters_with_issues",
          "issues_by_code",
          "meters",
          "network",
          "readings"
        ],
        "title": "DataQualityReport"
      },
...
      "MeterStatus": {
        "type": "string",
        "enum": [
          "installed",
          "faulty",
          "decommissioned",
          "unknown"
        ],
        "title": "MeterStatus"
      },
...
      "NetworkLevel": {
        "type": "string",
        "enum": [
          "zone",
          "circle",
          "division",
          "subdivision",
          "substation",
          "feeder",
          "transformer"
        ],
        "title": "NetworkLevel",
        "description": "Levels of the distribution network, top to bottom."
      },
...
      "NetworkNode": {
        "properties": {
          "level": {
            "$ref": "#/components/schemas/NetworkLevel"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "transformer_count": {
            "type": "integer",
            "title": "Transformer Count"
          },
          "parent_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Parent Codes",
            "description": "Every parent this code was observed under: one for circles and transformers, usually several for the levels in between (see /v1/network)."
          },
          "meters_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/MeterStatus"
            },
            "type": "object",
            "title": "Meters By Status"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/NetworkNeighbour"
            },
            "type": "array",
            "title": "Parents"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/NetworkNeighbour"
            },
            "type": "array",
            "title": "Children"
          },
          "transformers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Transformers",
            "description": "Codes of the distribution transformers under this node. A transformer node lists itself; its details are at /v1/transformers/{code}."
          },
          "name_variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Name Variants",
            "description": "Other names the portal uses for this node (stale data)."
          }
        },
        "type": "object",
        "required": [
          "level",
          "code",
          "name",
          "meter_count",
          "transformer_count",
          "parent_codes",
          "meters_by_status",
          "parents",
          "children",
          "transformers"
        ],
        "title": "NetworkNode"
      },
...
      "NetworkNodeSummary": {
        "properties": {
          "level": {
            "$ref": "#/components/schemas/NetworkLevel"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "transformer_count": {
            "type": "integer",
            "title": "Transformer Count"
          },
          "parent_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Parent Codes",
            "description": "Every parent this code was observed under: one for circles and transformers, usually several for the levels in between (see /v1/network)."
          }
        },
        "type": "object",
        "required": [
          "level",
          "code",
          "name",
          "meter_count",
          "transformer_count",
          "parent_codes"
        ],
        "title": "NetworkNodeSummary"
      },
      "NetworkOverview": {
        "properties": {
          "is_tree": {
            "type": "boolean",
            "title": "Is Tree",
            "description": "False if any level has children appearing under several parents."
          },
          "node_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/NetworkLevel"
            },
            "type": "object",
            "title": "Node Counts"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/NetworkEdge"
            },
            "type": "array",
            "title": "Edges"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "is_tree",
          "node_counts",
          "edges",
          "note"
        ],
        "title": "NetworkOverview"
      },
...
      "NetworkTreeNode": {
        "properties": {
          "level": {
            "$ref": "#/components/schemas/NetworkLevel"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "transformer_count": {
            "type": "integer",
            "title": "Transformer Count"
          },
          "meters_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/MeterStatus"
            },
            "type": "object",
            "title": "Meters By Status"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/NetworkTreeNode"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "type": "object",
        "required": [
          "level",
          "code",
          "name",
          "meter_count",
          "transformer_count",
          "meters_by_status"
        ],
        "title": "NetworkTreeNode"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable machine-readable error code:\n\n* `unauthorized` (401): `X-API-Key` is missing or wrong (only when the server sets `URJA_API_KEY`).\n* `meter_not_found` (404): No meter with that id in the index, or the portal no longer has it.\n* `transformer_not_found` (404): No distribution transformer with that code.\n* `network_node_not_found` (404): No network node with that level and code.\n* `not_found` (404): No such path.\n* `method_not_allowed` (405): The path exists, but not for this HTTP method.\n* `http_error` (4xx): Any other HTTP-level error.\n* `validation_error` (422): A parameter is malformed, out of range or not recognised; `errors` says which.\n* `granularity_unavailable` (422): Hourly consumption asked of a meter that reports once a day.\n* `internal_error` (500): A bug in this service; the server log has the details.\n* `upstream_auth_failed` (502): This service could not log in to the portal.\n* `upstream_protocol_error` (502): The portal answered in a format this service does not understand.\n* `sync_rejected` (502): A sync returned data that looked broken; the previous snapshot is kept.\n* `index_not_ready` (503): The first sync with the portal hasn't finished; `Retry-After` gives the next attempt.\n* `upstream_unavailable` (503): The portal is down or too slow, and there is no cached copy to serve.\n* `upstream_rate_limited` (503): The portal kept answering 429, and there is no cached copy to serve.\n* `busy` (503): Too many requests are queued for the portal's rate budget.",
            "examples": [
              "meter_not_found"
            ]
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors",
            "description": "Field-level validation errors."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details."
      },
...
      "TransformerDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "DT-001"
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Malviya Nagar DT 1"
            ]
          },
          "feeder_code": {
            "type": "string",
            "title": "Feeder Code",
            "examples": [
              "F-001"
            ]
          },
          "capacity_kva": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Kva",
            "examples": [
              100
            ]
          },
          "network": {
            "$ref": "#/components/schemas/NetworkPath",
            "description": "The transformer's own position in the network (zone \u2192 feeder)."
          },
          "meter_count": {
            "type": "integer",
            "title": "Meter Count"
          },
          "meters_by_status": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/MeterStatus"
            },
            "type": "object",
            "title": "Meters By Status"
          },
          "name_variants": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Name Variants",
            "description": "Other names meters report for this transformer (stale data)."
          },
          "meters": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Meters",
            "description": "Meters connected to this transformer."
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "feeder_code",
          "network",
          "meter_count",
          "meters_by_status",
          "meters"
        ],
        "title": "TransformerDetail"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Required only if the server is configured with URJA_API_KEY.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}